프로젝트 구조 커밋 (#4) - 컨피규레이션 (커넥션 풀) 추가
