+++ src/main/java/com/moment/admin/model/AdminDAO.java
@@ -5,9 +5,7 @@ import com.moment.configuration.ConnectionPool;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
+
 
 public class AdminDAO {
 
@@ -42,25 +40,4 @@ public class AdminDAO {
         cp.disconnect(rs, pstmt, conn);
     }
 
-    public List<AdminDTO> getNoticeList() throws SQLException {
-
-        List<AdminDTO> list = new ArrayList<>();
-        connect();
-
-        pstmt = conn.prepareStatement(AdminSQL.SELECTION_NOTICE_LIST.getQuery());
-        rs = pstmt.executeQuery();
-
-        while (rs.next()) {
-            AdminDTO dto = new AdminDTO();
-            dto.setNoticeNo(rs.getInt("notice_no"));
-            dto.setNoticeTitle(rs.getString("notice_title"));
-            dto.setNoticeCont(rs.getString("notice_cont"));
-            dto.setNoticeDate(rs.getString("notice_date"));
-            list.add(dto);
-        }
-
-        disconnect(rs, pstmt, conn);
-
-        return list;
-    }
 }
